factored out tasklog tables to separate table dir
