+++ src/tasklogs/components/TaskLogHistoryModal.tsx
@@ -1,10 +1,11 @@
 import { useState } from "react"
 import Modal from "src/core/components/Modal"
 import Table from "src/core/components/Table"
-import { taskLogHistoryTableColumns, taskLogHistoryTableColumnsNoMeta } from "./TaskLogHistoryTable"
 import { Prisma } from "@prisma/client"
-import { processTaskLogHistory } from "src/tasklogs/utils/processTaskLogs"
+import { processTaskLogHistory } from "src/tasklogs/tabels/processing/processTaskLogs"
 import { ExtendedTaskLog } from "../hooks/useTaskLogData"
+import { TaskLogHistoryFormColumns } from "../tabels/columns/TaskLogHistoryFormColumns"
+import { TaskLogHistoryCompleteColumns } from "../tabels/columns/TaskLogHistoryCompleteColumns"
 
 type AssignmentHistoryModalProps = {
   taskLogs: ExtendedTaskLog[]
@@ -20,8 +21,6 @@ export const AssignmentHistoryModal = ({ taskLogs, schema, ui }: AssignmentHisto
 
   const processedAssignmentHistory = processTaskLogHistory(taskLogs, schema, ui)
 
-  //console.log(taskLogs)
-
   return (
     <>
       <div className="">
@@ -33,7 +32,7 @@ export const AssignmentHistoryModal = ({ taskLogs, schema, ui }: AssignmentHisto
           <h1 className="flex justify-center mb-2 text-3xl">Task History</h1>
           <div className="modal-action flex flex-col">
             <Table
-              columns={schema && ui ? taskLogHistoryTableColumns : taskLogHistoryTableColumnsNoMeta}
+              columns={schema && ui ? TaskLogHistoryFormColumns : TaskLogHistoryCompleteColumns}
               data={processedAssignmentHistory}
               classNames={{
                 thead: "text-base",
