+++ src/tasklogs/tabels/columns/TaskLogFormColumns.tsx
@@ -1,0 +1,36 @@
+import React from "react"
+import { ColumnDef, createColumnHelper } from "@tanstack/react-table"
+import { ProcessedIndividualTaskLog } from "../processing/processTaskLogs"
+import { TaskLogSchemaModal } from "../../components/TaskLogSchemaModal"
+
+// Column helper
+const columnHelper = createColumnHelper<ProcessedIndividualTaskLog>()
+
+// ColumnDefs
+// Table for assignment with a form
+export const TaskLogFormColumns: ColumnDef<ProcessedIndividualTaskLog>[] = [
+  columnHelper.accessor("projectMemberName", {
+    cell: (info) => <span>{`${info.getValue()}`}</span>,
+    header: "Contributor Name",
+  }),
+  columnHelper.accessor("lastUpdate", {
+    cell: (info) => <span>{info.getValue()}</span>,
+    header: "Last Update",
+    id: "updatedAt",
+  }),
+  columnHelper.accessor("status", {
+    cell: (info) => <span>{info.getValue()}</span>,
+    header: "Status",
+    id: "status",
+  }),
+  columnHelper.accessor("taskLog", {
+    cell: (info) => {
+      return (
+        <>
+          <TaskLogSchemaModal taskLog={info.getValue()} />
+        </>
+      )
+    },
+    header: "Form Data",
+  }),
+]
