+++ src/tasklogs/tabels/columns/TeamTaskLogCompleteColumns.tsx
@@ -1,15 +1,14 @@
 import React from "react"
 import { ColumnDef, createColumnHelper } from "@tanstack/react-table"
-import { TaskLogToggleModal } from "./TaskLogToggleModal"
-import { ProcessedTeamTaskLog } from "../utils/processTaskLogs"
-import { TaskLogSchemaModal } from "./TaskLogSchemaModal"
+import { TaskLogToggleModal } from "../../components/TaskLogToggleModal"
+import { ProcessedTeamTaskLog } from "../processing/processTaskLogs"
 import { ShowTeamModal } from "src/teams/components/ShowTeamModal"
 
 // Column helper
 const columnHelper = createColumnHelper<ProcessedTeamTaskLog>()
 
 // ColumnDefs
-export const teamTaskLogTableColumns: ColumnDef<ProcessedTeamTaskLog>[] = [
+export const TeamTaskLogCompleteColumns: ColumnDef<ProcessedTeamTaskLog>[] = [
   columnHelper.accessor("projectMember", {
     cell: (info) => {
       const rowData = info.row.original // Access the entire row data
@@ -41,32 +40,5 @@ export const teamTaskLogTableColumns: ColumnDef<ProcessedTeamTaskLog>[] = [
     },
     header: "Change status",
     id: "updateStatus",
-  }),
-]
-
-export const teamTaskLogTableColumnsSchema: ColumnDef<ProcessedTeamTaskLog>[] = [
-  columnHelper.accessor("projectMember", {
-    cell: (info) => <div>{<ShowTeamModal projectMember={info.getValue()}></ShowTeamModal>}</div>,
-    header: "Team Name",
-  }),
-  columnHelper.accessor("lastUpdate", {
-    cell: (info) => <span>{info.getValue()}</span>,
-    header: "Last Update",
-    id: "updatedAt",
-  }),
-  columnHelper.accessor("status", {
-    cell: (info) => <span>{info.getValue()}</span>,
-    header: "Status",
-    id: "status",
-  }),
-  columnHelper.accessor("taskLog", {
-    cell: (info) => {
-      return (
-        <>
-          <TaskLogSchemaModal taskLog={info.getValue()} />
-        </>
-      )
-    },
-    header: "Form Data",
   }),
 ]
